feat(archive/tests)✨: Add tests for SQLiteVecDocStore
- Create a new test file for SQLiteVecDocStore functionalities.
- Include tests for document addition, retrieval, and reset operations.
